+++ Lab3_Estiven_Maikol/src/ProtectedAreasList.java
@@ -41,7 +41,7 @@ public class ProtectedAreasList extends ProtectedAreas {
     }
 
     @Override
-    public int income() {
+    public double income() {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
